feat(deck-layers): scaffold @cosmos.gl/deck-layers and port the prototype layers
The zero-copy deck.gl layers lived as story-local prototype code, unusable
outside this repo's Storybook. They become the first workspace package:
@cosmos.gl/deck-layers at integrations/deck-layers, version-locked to the
root package, so hosts can install the texture-sampling layers instead of
copying them from an example.

- The package peers on @cosmos.gl/graph (workspace:^, published as a real
  range), @deck.gl/core and @luma.gl/core+engine — everything stays
  external in its ES build because the host must resolve one luma
  installation and owns the simulation the layers sample. The catalog in
  pnpm-workspace.yaml becomes the single home of the deck/luma dev pins
  the root and the package must agree on.
- vite-plugin-dts gets aliasesExclude for @cosmos.gl/graph: the package
  builds against live root source through a path alias, but emitted
  declarations must keep importing the package name, not reach into
  ../../src.
- The root maps @cosmos.gl/deck-layers to the package source in
  vite.config.ts (which Storybook auto-merges) and tsconfig paths — an
  alias instead of a root devDependency, so the workspace stays free of a
  root↔package dependency cycle.
- The package tsconfig includes ../../src/declaration.d.ts: root sources
  arrive through the path alias, but their ambient *?raw shader-module
  declarations only arrive via include.
- The zero-copy story now imports the package by name and shows the two
  layer sources as separate panes; the story-local prototype file is
  gone. Root lint/typecheck/build chain into the workspace via pnpm -r.

Verified: lint, typecheck (root + package), build (root ES+UMD + package,
no 'error TS'; package d.ts imports stay on '@cosmos.gl/graph', bundle
keeps deck/luma external), build:storybook, and the 22 browser tests.
