+++ .eslintrc
@@ -15,9 +15,9 @@
   ],
   "parser": "@typescript-eslint/parser",
   "parserOptions": {
-    "project": ["./tsconfig.json", "./src/stories/tsconfig.json", "./test/tsconfig.json"]
+    "project": ["./tsconfig.json", "./src/stories/tsconfig.json", "./test/tsconfig.json", "./integrations/deck-layers/tsconfig.typecheck.json"]
   },
-  "ignorePatterns": ["vite.config.ts", "vitest.config.ts"],
+  "ignorePatterns": ["**/vite.config.ts", "vitest.config.ts"],
   "rules": {
     "object-curly-spacing": ["error", "always"],
     "prefer-template": "error",
