+++ integrations/deck-layers/src/blend-parameters.ts
@@ -1,0 +1,15 @@
+/**
+ * Premultiplied-style alpha blending with depth writes disabled: points and
+ * links are drawn back-to-front by layer order, never by depth.
+ */
+export const BLEND_PARAMETERS = {
+  depthWriteEnabled: false,
+  depthCompare: 'always',
+  blend: true,
+  blendColorOperation: 'add',
+  blendColorSrcFactor: 'src-alpha',
+  blendColorDstFactor: 'one-minus-src-alpha',
+  blendAlphaOperation: 'add',
+  blendAlphaSrcFactor: 'one',
+  blendAlphaDstFactor: 'one-minus-src-alpha',
+} as const
