+++ integrations/deck-layers/tsconfig.typecheck.json
@@ -1,0 +1,17 @@
+{
+  // Whole-package typecheck (`pnpm run typecheck`): everything under src,
+  // stories included, with no emit.
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "noEmit": true,
+    "declaration": false
+  },
+  "include": [
+    "./src/**/*",
+    "../../src/declaration.d.ts"
+  ],
+  "exclude": [
+    "node_modules",
+    "dist"
+  ]
+}
