+++ package.json
@@ -14,12 +14,12 @@
   ],
   "scripts": {
     "prepare": "sh scripts/prepare.sh",
-    "build": "vite build --mode es && vite build --mode umd",
+    "build": "vite build --mode es && vite build --mode umd && pnpm -r run build",
     "watch": "vite build --watch",
     "test": "vitest run",
     "test:watch": "vitest",
-    "typecheck": "tsc -p tsconfig.typecheck.json",
-    "lint": "eslint --cache ./src ./test --ext .ts --ignore-path .gitignore",
+    "typecheck": "tsc -p tsconfig.typecheck.json && pnpm -r run typecheck",
+    "lint": "eslint --cache ./src ./test ./integrations --ext .ts --ignore-path .gitignore",
     "lint:staged": "pnpm exec lint-staged",
     "storybook": "storybook dev -p 6006",
     "build:storybook": "storybook build",
@@ -39,14 +39,14 @@
   "homepage": "https://cosmosgl.github.io/graph",
   "author": "cosmos.gl",
   "devDependencies": {
-    "@deck.gl/core": "~9.3.0",
-    "@deck.gl/layers": "~9.3.0",
+    "@deck.gl/core": "catalog:",
+    "@deck.gl/layers": "catalog:",
     "@interacta/css-labels": "^0.6.0",
     "@jls-digital/storybook-addon-code": "^1.0.4",
-    "@luma.gl/core": "~9.3.6",
-    "@luma.gl/engine": "~9.3.6",
-    "@luma.gl/shadertools": "~9.3.6",
-    "@luma.gl/webgl": "~9.3.6",
+    "@luma.gl/core": "catalog:",
+    "@luma.gl/engine": "catalog:",
+    "@luma.gl/shadertools": "catalog:",
+    "@luma.gl/webgl": "catalog:",
     "@storybook/addon-docs": "^8.4.5",
     "@storybook/addon-essentials": "^8.4.5",
     "@storybook/addon-interactions": "^8.4.5",
@@ -81,9 +81,9 @@
     "playwright": "^1.62.1",
     "remark-gfm": "^4.0.0",
     "storybook": "^8.4.5",
-    "typescript": "^5.9.3",
-    "vite": "^6.4.3",
-    "vite-plugin-dts": "^4.3.0",
+    "typescript": "catalog:",
+    "vite": "catalog:",
+    "vite-plugin-dts": "catalog:",
     "vitest": "^4.1.11"
   },
   "peerDependencies": {
