+++ tsconfig.json
@@ -6,6 +6,8 @@
       "@/graph/*": ["src/*"],
       "@cosmos.gl/graph": ["src/"],
       "@cosmos.gl/graph/*": ["src/*"],
+      "@cosmos.gl/deck-layers": ["integrations/deck-layers/src/"],
+      "@cosmos.gl/deck-layers/*": ["integrations/deck-layers/src/*"],
     },
     "target": "es2019",
     "module": "esnext",
